Additional clean up and add more getter and setter functions
